requirements.txt
absl-py
attrs
cffi
contourpy
cycler
flatbuffers
fonttools
jax
jaxlib
kiwisolver
matplotlib
mediapipe
ml_dtypes
numpy
opencv-contrib-python
opencv-python
opt_einsum
packaging
pillow
protobuf
pycparser
pyparsing
python-dateutil
scipy
six
sounddevice
google-generativeai
python-dotenv